prometheus: drop `monitor` label from alerts
This was a long-standing bug: the `monitor` label is added to all metrics to
allow tracking which prometheus instance scraped a metric in a federated/HA
set-up, but then the AlertManager cannot de-duplicate alerts coming from
different prometheus instances as the values are different.

This change drops the label before the alerts are sent to the AlertManager.